+++ CHANGELOG.md
@@ -2,10 +2,10 @@
 
 ## 2.0 (work in progress)
 
-- Open any URL in header and footer
 - Add `initWithAcknowledgements:` initializer on `VTAcknowledgementsViewController`
 - Add `initWithAcknowledgements:style` initializer on `VTAcknowledgementsViewController`
 - Add `initWithPath:style` initializer on `VTAcknowledgementsViewController`
+- Open any URL in header and footer
 - Update `UIViewController` subclasses initializers nullability annotations to match UIKit methods
 - Update deployment target to iOS 9
 - Refactor localization to dedicated class